+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/servlets/action/UpdateMessageEntryServlet.java
@@ -2,6 +2,7 @@ package be.orbinson.aem.dictionarytranslator.servlets.action;
 
 import be.orbinson.aem.dictionarytranslator.exception.DictionaryException;
 import be.orbinson.aem.dictionarytranslator.services.impl.CombiningMessageEntryResourceProvider;
+import be.orbinson.aem.dictionarytranslator.servlets.datasource.CombiningMessageEntryDatasourceForDialog;
 import be.orbinson.aem.dictionarytranslator.services.DictionaryService;
 import be.orbinson.aem.dictionarytranslator.services.Dictionary;
 
@@ -20,11 +21,15 @@ import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.day.cq.replication.ReplicationException;
+import com.day.cq.replication.Replicator;
+
 import javax.jcr.RepositoryException;
 import javax.servlet.Servlet;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Locale;
+import java.util.Optional;
 
 @Component(service = Servlet.class)
 @SlingServletResourceTypes(
@@ -38,6 +43,9 @@ public class UpdateMessageEntryServlet extends AbstractDictionaryServlet {
 
     @Reference
     private transient DictionaryService dictionaryService;
+
+    @Reference
+    private Replicator replicator;
 
     @Override
     protected void doPost(SlingHttpServletRequest request, @NotNull SlingHttpServletResponse response) throws IOException {
@@ -63,17 +71,25 @@ public class UpdateMessageEntryServlet extends AbstractDictionaryServlet {
         }
     }
 
-    private void updateCombiningMessageEntry(SlingHttpServletRequest request, ResourceResolver resourceResolver, Resource combiningMessageEntryResource) throws DictionaryException, PersistenceException, RepositoryException {
-        String key = getMandatoryParameter(request, "key", false);
+    private void updateCombiningMessageEntry(SlingHttpServletRequest request, ResourceResolver resourceResolver, Resource combiningMessageEntryResource) throws DictionaryException, PersistenceException, RepositoryException, ReplicationException {
         String dictionaryPath = combiningMessageEntryResource.getValueMap().get(CombiningMessageEntryResourceProvider.DICTIONARY_PATH, String.class);
+        String key = combiningMessageEntryResource.getValueMap().get(CombiningMessageEntryResourceProvider.KEY, String.class);
         if (StringUtils.isNotBlank(dictionaryPath)) {
             // check languages
             Locale[] languages = combiningMessageEntryResource.getValueMap().get(CombiningMessageEntryResourceProvider.LANGUAGES, new Locale[0]);
             for (Locale language : languages) {
-                String message = getMandatoryParameter(request, language.toLanguageTag(), true); // ensure the parameter exists
                 Dictionary dictionary = dictionaryService.getDictionary(resourceResolver, dictionaryPath, language)
                         .orElseThrow(() -> new DictionaryException("Could not find dictionary for language '" + language + "' below path: " + dictionaryPath));
-                dictionary.createOrUpdateEntry(resourceResolver, key, message);
+                String paramNameMessage = language.toLanguageTag();
+                String message = getOptionalParameter(request, paramNameMessage, true).orElse("");
+                boolean useEmpty = getOptionalParameter(request, paramNameMessage + CombiningMessageEntryDatasourceForDialog.SUFFIX_USE_EMPTY, false, Boolean::parseBoolean).orElse(false);
+                final Optional<String> messageOptional;
+                if (!useEmpty && message.isEmpty()) {
+                    messageOptional = Optional.empty();
+                } else {
+                    messageOptional = Optional.of(message);
+                }
+                dictionary.updateEntry(resourceResolver, key, messageOptional);
             }
             resourceResolver.commit();
         } else {