+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/servlets/datasource/LanguageDatasource.java
@@ -1,7 +1,7 @@
 package be.orbinson.aem.dictionarytranslator.servlets.datasource;
 
 import java.io.IOException;
-import java.text.Collator;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -19,7 +19,6 @@ import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
 import org.apache.sling.api.request.RequestDispatcherOptions;
 import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.api.servlets.SlingSafeMethodsServlet;
 import org.apache.sling.api.wrappers.ValueMapDecorator;
@@ -33,10 +32,11 @@ import org.slf4j.LoggerFactory;
 import com.adobe.granite.ui.components.Config;
 import com.adobe.granite.ui.components.ds.DataSource;
 import com.adobe.granite.ui.components.ds.SimpleDataSource;
-import com.adobe.granite.ui.components.ds.ValueMapResource;
 
 import be.orbinson.aem.dictionarytranslator.services.DictionaryService;
 import be.orbinson.aem.dictionarytranslator.services.impl.DictionaryImpl;
+import be.orbinson.aem.dictionarytranslator.servlets.datasource.builder.ValueMapResourceBuilderFactory;
+import be.orbinson.aem.dictionarytranslator.servlets.datasource.builder.graniteui.ComponentValueMapResourceBuilder;
 
 @Component(service = Servlet.class)
 @SlingServletResourceTypes(
@@ -51,6 +51,26 @@ public class LanguageDatasource extends SlingSafeMethodsServlet {
     @Reference
     transient DictionaryService dictionaryService;
 
+    private static class GraniteFormSelectItem {
+        private final ValueMap properties;
+        private GraniteFormSelectItem(ValueMap properties) {
+            this.properties = properties;
+        }
+
+        public static GraniteFormSelectItem fromResource(Resource resource) {
+            return new GraniteFormSelectItem(resource.getValueMap());
+        }
+
+        public String getText() {
+            return properties.get("text", String.class);
+        }
+
+        public String getValue() {
+            return properties.get("value", String.class);
+        }
+
+    }
+
     public static Map<Locale, String> getAllAvailableLanguages(SlingHttpServletRequest request, @NotNull SlingHttpServletResponse response) throws ServletException, IOException {
         // TranslationConfig.getLanguages(ResourceResolver) does never return the country label,
         // therefore use the data source which is also used in the Page Properties dialog (Advanced Tab in Language)
@@ -58,11 +78,11 @@ public class LanguageDatasource extends SlingSafeMethodsServlet {
         options.setForceResourceType("cq/gui/components/common/datasources/languages");
         request.getRequestDispatcher(request.getResource(), options).include(request, response);
         DataSource dataSource = (DataSource) request.getAttribute(DataSource.class.getName());
-        List<ValueTextResource> resources = IteratorUtils.toList(new TransformIterator<>(dataSource.iterator(), r -> ValueTextResource.fromResource(request.getLocale(), r)));
+        List<GraniteFormSelectItem> resources = IteratorUtils.toList(new TransformIterator<>(dataSource.iterator(), r -> GraniteFormSelectItem.fromResource(r)));
         return toLanguageMap(resources);
     }
 
-    private static Map<Locale, String> toLanguageMap(List<ValueTextResource> resources) {
+    private static Map<Locale, String> toLanguageMap(List<GraniteFormSelectItem> resources) {
         return resources.stream()
                 // the upstream data source does not filter access control child resource
                 .filter(r -> !AccessControlConstants.REP_POLICY.equals(r.getValue()))
@@ -92,86 +112,30 @@ public class LanguageDatasource extends SlingSafeMethodsServlet {
         } else {
             languageFilter = l -> !dictionaryLanguages.contains(l);
         }
-        // convert to list of resources
-        boolean emitTextFieldResources = dsCfg.get("emitTextFieldResources", false);
-        List<OrderedValueMapResource> resourceList = languageMap.entrySet().stream()
-                .filter(e -> languageFilter.test(e.getKey()))
-                .map(e -> {
-                    if (emitTextFieldResources) {
-                        return TextFieldResource.create(request.getLocale(), request.getResourceResolver(), e.getKey().toLanguageTag(), e.getValue());
-                    } else {
-                        return ValueTextResource.create(request.getLocale(), request.getResourceResolver(), e.getKey().toLanguageTag(), e.getValue());
-                    }
-                }).sorted().collect(Collectors.toList());
-        // sort by display names
+        final List<Resource> resources;
+        boolean forCreateEntryDialog = dsCfg.get("forCreateEntryDialog", false);
+        if (forCreateEntryDialog) {
+            // populate the create entry dialog
+            Locale[] languages = languageMap.entrySet().stream()
+                    .filter(e -> languageFilter.test(e.getKey()))
+                    .map(Map.Entry::getKey)
+                    .toArray(Locale[]::new);
+            resources = new LinkedList<>();
+            CombiningMessageEntryDatasourceForDialog.createCombiningMessageEntryDataSource(request.getLocale(), languageMap, request.getResourceResolver(), resources, new ValueMapDecorator(Map.of()), languages);
+        } else {
+            ValueMapResourceBuilderFactory factory = new ValueMapResourceBuilderFactory(request.getResourceResolver(), "/languages");
+            // otherwise just populate a select control (granite/ui/components/coral/foundation/form/select)
+            resources = languageMap.entrySet().stream()
+                    .filter(e -> languageFilter.test(e.getKey()))
+                    .map(e -> {
+                        return ComponentValueMapResourceBuilder.forFormSelectItem(factory, e.getValue(), e.getKey().toLanguageTag()).build();
+                    }).collect(Collectors.toList());
+            // sort by language label
+            CombiningMessageEntryDatasourceForDialog.sortResourcesByProperty("text", request.getLocale(), resources);
+        }
+        
         // create data source (only accepts iterator over Resource, not of subclasses so we need to transform)
-        DataSource dataSource = new SimpleDataSource(new TransformIterator<>(resourceList.iterator(), r -> (Resource) r));
+        DataSource dataSource = new SimpleDataSource(new TransformIterator<>(resources.iterator(), r -> (Resource) r));
         request.setAttribute(DataSource.class.getName(), dataSource);
-    }
-
-    private abstract static class OrderedValueMapResource extends ValueMapResource implements Comparable<OrderedValueMapResource> {
-
-        private final Collator collator;
-
-        protected OrderedValueMapResource(Locale locale, ResourceResolver resourceResolver, String resourceType, ValueMap vm) {
-            super(resourceResolver, "", resourceType, vm);
-            collator = Collator.getInstance(locale);
-        }
-
-        abstract String getLabel();
-
-        @Override
-        public int compareTo(OrderedValueMapResource o) {
-            return collator.compare(getLabel(), o.getLabel());
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            return obj instanceof OrderedValueMapResource && collator.equals(((OrderedValueMapResource) obj).collator);
-        }
-    }
-
-    private static class TextFieldResource extends OrderedValueMapResource {
-
-        private TextFieldResource(Locale locale, ResourceResolver resolver, ValueMap valueMap) {
-            super(locale, resolver, "granite/ui/components/coral/foundation/form/textfield", valueMap);
-        }
-
-        public static TextFieldResource create(Locale locale, ResourceResolver resolver, String value, String text) {
-            ValueMap valueMap = new ValueMapDecorator(Map.of("fieldLabel", text, "name", value));
-            return new TextFieldResource(locale, resolver, valueMap);
-        }
-
-        String getLabel() {
-            return getValueMap().get("fieldLabel", String.class);
-        }
-    }
-
-    private static class ValueTextResource extends OrderedValueMapResource {
-        private ValueTextResource(Locale locale, ResourceResolver resolver, ValueMap valueMap) {
-            super(locale, resolver, "nt:unstructured", valueMap);
-        }
-
-        public static ValueTextResource fromResource(Locale locale, Resource resource) {
-            return new ValueTextResource(locale, resource.getResourceResolver(), resource.getValueMap());
-        }
-
-        public static ValueTextResource create(Locale locale, ResourceResolver resolver, String value, String text) {
-            ValueMap valueMap = new ValueMapDecorator(Map.of("value", value, "text", text));
-            return new ValueTextResource(locale, resolver, valueMap);
-        }
-
-        public String getText() {
-            return getValueMap().get("text", String.class);
-        }
-
-        public String getValue() {
-            return getValueMap().get("value", String.class);
-        }
-
-        @Override
-        String getLabel() {
-            return getText();
-        }
     }
 }