+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/services/impl/DictionaryServiceImplTest.java
@@ -1,9 +1,9 @@
 package be.orbinson.aem.dictionarytranslator.services.impl;
 
-import static org.junit.Assert.assertNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertIterableEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;